% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_clean_names.R
\name{cl_clean_names}
\alias{cl_clean_names}
\title{Clean data.table names to be lowercase and separated by a period
\code{.}}
\usage{
cl_clean_names(DT)
}
\arguments{
\item{DT}{a \code{data.table}}
}
\description{
This function uses the \code{janitor::make_clean_names} function,
but replaces underscores with a period. Note that this function
updates the \code{data.table} by reference
}
\examples{
DT <- data.table(`Var A` = 1, VarB = 2, var_c = 3)
print(cl_clean_names(DT))
}
