% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_readxl_backslashes.R
\name{clean_readxl_backslashes}
\alias{clean_readxl_backslashes}
\title{Clean extra backslashes from readxl}
\usage{
clean_readxl_backslashes(string)
}
\arguments{
\item{string}{a character string}
}
\value{
A character string with the extra backslashes inserted by
  readxl removed.
}
\description{
\code{clean_readxl_backslashes()} cleans extras backslashes that
readxl uses as escape characters when readxl reads in data. More
specifically, readxl converts "\n" to "\\n", "\u00ad" to "\\u00ad",
etc.
}
\examples{
## https://gist.github.com/ChandlerLutz/cca06fb35003713efae87c0322ac193e
## New line 
clean_readxl_backslashes("\\\\n")
## soft hyphen
clean_readxl_backslashes("\\\\u00ad")
## en dash
clean_readxl_backslashes("\\\\u2013")
## em dash
clean_readxl_backslashes("\\\\u2014")
}
