% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLmisc_main.R
\docType{package}
\name{CLmisc-package}
\alias{CLmisc}
\alias{CLmisc-package}
\title{Miscellaneous Functions by Chandler Lutz}
\description{
Chandler Lutz's Miscellaneous R functions.
}
\author{
\strong{Maintainer}: Chandler Lutz \email{chandler.lutz@gmail.com}

}
