% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exported.R
\name{convert_file_path_to_windows_format}
\alias{convert_file_path_to_windows_format}
\title{Convert a file path to windows format}
\usage{
convert_file_path_to_windows_format(.file.path)
}
\arguments{
\item{.file.path}{a unix style file path}
}
\description{
Convert a file path to windows format by turning forward slashes
`/` to backslashes `\`
}
\examples{
print(getwd())
convert_file_path_to_windows_format(getwd()) \%>\% print
}
