% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_trends.R
\name{create_time_trends}
\alias{create_time_trends}
\title{Fast creatation of time trends}
\usage{
create_time_trends(
  .data,
  time.col,
  select.column,
  quadratic.time.trends = FALSE,
  remove.first.trend = FALSE
)
}
\arguments{
\item{.data}{a \code{data.table}}

\item{time.col}{a time column; can be of any type. `time.col` will
be ordered using \code{order()}}

\item{select.column}{a column with cross-sectional units that will
be used to create the time trends. Think fixed effects such as
cities, regions, counties, etc.}

\item{quadratic.time.trends}{a \code{logical} for whether to
include quadratic (squared) time trends __IN ADDITION__ to the
linear time trends for the the dummy columns. Default is
\code{FALSE}}

\item{remove.first.trend}{removes the first time trend so that only
\code{N - 1} time trends reamin. This avoids the
multicollinearity in models. Default is \code{FALSE}}
}
\value{
A data.table with the same number of rows, but the time
  trends included. The format of the time trend columns will be
  \code{tt.select.column_value} (linear trends) or
  \code{tt2.select.column_value} (quadratic trends) where
  \code{select.column} is the name of the character column
  requested from the user and \code{value} is the value of the of
  the character column for that time trend
}
\description{
create_time_trends() quickly creates time trends based on a
character columns from inputted data. This function is useful for
statistical analysis when you want time trends. The function will
only operate on `data.tables` by reference.
}
\examples{
crime <- setDT(expand.grid(city = c("SF", "SF", "NYC"),
               year = c(1990, 2000, 2010),
               crime = 1:3, stringsAsFactors = FALSE))
crime2 <- copy(crime)
print(create_time_trends(crime2, time.col = "year",
                         select.column = "city"))
}
\seealso{
\code{\link{dummy_cols}} For creating dummy columns

Other dummy functions: 
\code{\link{dummy_cols}()},
\code{\link{dummy_rows}()}
}
\concept{dummy functions}
