% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_all_packages.R
\name{detach_all_packages}
\alias{detach_all_packages}
\title{Detach all but base packages in R}
\usage{
detach_all_packages()
}
\description{
From https://stackoverflow.com/a/8817519/1317443
}
