% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fntile.R
\name{fntile}
\alias{fntile}
\title{fast ntile}
\usage{
fntile(x, n)
}
\arguments{
\item{x}{vector of data}

\item{n}{the numbe of quantiles desired}
}
\value{
A vector corresponding to the ntiles. (As in \code{dplyr::ntile}.)
}
\description{
fast ntile
}
\examples{
fntile(c(NA, 1:20, NA), 5)
}
