% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_limits.R
\name{get_index_limits}
\alias{get_index_limits}
\title{Convenient Function to get the index limits of a \code{data.frame} or data.table}
\usage{
get_index_limits(.data, min.extend = 0, max.extend = 0)
}
\arguments{
\item{.data}{a \code{data.table} with a variable \code{index}}

\item{min.extend}{how much to extend the limits relative to the
minimum}

\item{max.extend}{how much to extend the limits relative to the
maximum}
}
\value{
a numeric vector of length 2, where the first element is
  the minimum and the second element is the maximum
}
\description{
Convenient function to get the index limits of a \code{data.table} with a variable called \code{index}. This function is especially useful to extend the axis limits of a `ggplot` using the \code{min.extend} and \code{max.extend} arguments of this function.
}
\examples{
DT <- data.table(index = as.Date(c("2010-01-01", "2015-01-01", "2021-01-01")))
get_index_limits(DT)
get_index_limits(DT, max.extend = 365)
get_index_limits(DT, min.extend = 365)
get_index_limits(DT, min.extend = -365)
}
