% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rtigrs.R
\name{get_rtigris}
\alias{get_rtigris}
\title{Retrieve and cache data from the tigris package.}
\usage{
get_rtigris(tigris_fun, save_dir = here::here("data-raw/rtigris/"), ...)
}
\arguments{
\item{tigris_fun}{A function from the \code{\link[tigris:tigris-package]{tigris}} package.  This should be passed as a bare function name (e.g., \code{zctas}, not \code{"zctas"}).}

\item{save_dir}{The directory where the cached RDS files should be saved. Defaults to \code{here::here("data-raw/tigris/")}.}

\item{...}{Additional arguments to be passed to the \code{tigris_fun}.  These arguments will be incorporated into the name of the cached file, allowing for different cached versions based on different function arguments.}
}
\value{
An sf object containing the data retrieved from the specified \code{tigris_fun}.
}
\description{
This function retrieves data from the \code{\link[tigris:tigris-package]{tigris}} package, caches the results as an RDS file, and returns the data.  It avoids redundant downloads by checking for the existence of a cached file before attempting to download.
}
\details{
This function simplifies the process of working with the \code{tigris} package by automatically caching downloaded data.  It constructs a file name based on the function called and its arguments, ensuring that different datasets or different queries to the same dataset are stored separately. This prevents unnecessary downloads and speeds up subsequent access to the same data.

The function performs several checks to ensure that the provided `tigris_fun` is a valid function from the `tigris` package and that the returned object is an sf object with a geometry column.
}
\examples{
\dontrun{
# Get ZCTAs for California and cache the result.
ca_zctas <- get_rtigris(zctas, state = "CA")

# Get ZCTAs for California with a different year and cache the result separately.
ca_zctas_2020 <- get_rtigris(zctas, state = "CA", year = 2020)

# Get counties for North Carolina.
nc_counties <- get_rtigris(counties, state = "NC")
}

}
