% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_breaks.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Create integer breaks for ggplot2 axes}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{The number of breaks}

\item{...}{extra parameters passes to \code{pretty}}
}
\description{
See https://stackoverflow.com/a/62321155
}
\examples{
data(mtcars)
ggplot(mtcars[mtcars$wt > 3, ], aes(x = cyl, y = wt)) +
 geom_point() +
 scale_y_continuous(breaks = integer_breaks())
}
