% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages_in_folder.R
\name{list_packages_in_folder}
\alias{list_packages_in_folder}
\title{List all of the packages used in a folder of R scripts}
\usage{
list_packages_in_folder(folder)
}
\arguments{
\item{folder}{string with the path to the folder}
}
\value{
character vector with the packages
}
\description{
Searches through a folder, fitlers out the R scripts and then
returns a character vector with the R packages. Note: this function
will not list package dependencies
}
