% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exported.R
\name{mkdir_p}
\alias{mkdir_p}
\title{Recursively make a directory if it does not exist}
\usage{
mkdir_p(dir)
}
\arguments{
\item{dir}{the directory (including parent directories) to create}
}
\description{
Similar to the linux command \code{mkir -p}
}
\examples{
\donttest{
mkdir_p("test/test")
}
}
