% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_by_ref.R
\name{mutate_by_ref}
\alias{mutate_by_ref}
\title{Mutate multiple cols by reference using the same function}
\usage{
mutate_by_ref(DT, cols, f, names.extra = NULL)
}
\arguments{
\item{DT}{a \code{data.table}}

\item{cols}{a character vector of column names or a function that
returns a logical vector to choose the column names}

\item{f}{a function that will be used to mutate the columns}

\item{names.extra}{if \code{NULL} (the default), then \code{cols}
will be updated by reference in plcace. If \code{names.extra} is
a character string, then new columns will be appended to
\code{DT} with the names formed by \code{paste0(var,
names.extra)}. If \code{names.extra} is set to \code{TRUE}, then
new columns will be appended to \code{DT} with the names
\code{paste0(var, ".", f)}, where \code{f} is the name of the
function passed by the user and is found using
\code{deparse(substitute(f))}}
}
\description{
Mutate multiple cols by reference using the same function
}
\examples{
data(mtcars)

DT.mtcars <- as.data.table(mtcars)

##Add some columns for testing
DT.mtcars <- DT.mtcars \%>\%
  .[, mpg.char := as.character(mpg)] \%>\%
  .[, cyl.fact := as.factor(cyl)]

## mutate in place by reference
mutate_by_ref(copy(DT.mtcars),
              cols = c("mpg", "hp"), f = sqrt) \%>\%
  head \%>\% print

## mutate with new column names by reference
mutate_by_ref(copy(DT.mtcars),
              cols = c("mpg", "hp"), f = sqrt, names.extra = ".sqrt") \%>\%
  head \%>\% print

## mutate with new column names by reference
## setting `names.extra` to TRUE will use `deparse(substitute(f))` to
## get the name of f to use as `names.extra` and sepearte with a dot (`.`)
mutate_by_ref(copy(DT.mtcars),
              cols = c("mpg", "hp"), f = sqrt, names.extra = TRUE) \%>\%
  head \%>\% print

##mutate all numeric columns in place by reference
mutate_by_ref(copy(DT.mtcars),
              cols = is.numeric, f = sqrt) \%>\%
  head \%>\% print

##mutate all numeric columns with new names by reference
mutate_by_ref(copy(DT.mtcars),
              cols = is.numeric, f = sqrt, names.extra = ".sqrt") \%>\%
  head \%>\% print

##mutate all numeric columns with new names by reference
##setting `names.extra` to TRUE
mutate_by_ref(copy(DT.mtcars),
              cols = is.numeric, f = sqrt, names.extra = TRUE) \%>\%
  head \%>\% print
}
