% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_and_close_excel.R
\name{open_and_close_excel}
\alias{open_and_close_excel}
\title{Open and Close an excel file}
\usage{
open_and_close_excel(.file.path, time.sleep = 30)
}
\arguments{
\item{.file.path}{A path to the excel file to open and close}

\item{time.sleep}{The number of seconds to leave the excel file
open. Defaults to 30 seconds}
}
\description{
Using Visual Basic to open and close an excel file. Must be on
windows. The script will create a temporary visual basic file
`temp__.vbs`
}
