% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_test.R
\name{pipe_test}
\alias{pipe_test}
\title{Test in a \code{magrittr} pipe}
\usage{
pipe_test(obj, test, value, tolerance = NULL)
}
\arguments{
\item{obj}{An \code{R} object. When used in a \code{magrittr}, this
will be omitted. This will also be returned}

\item{test}{The test}

\item{value}{the value that the test should be equal to}

\item{tolereance}{The tolerance for equality. Set to 0 exact
(identical) equality}
}
\value{
\code{obj} -- the original first parameter passed to
  \code{pipe_test}
}
\description{
To run a test in \code{magrittr} pipeline
}
\examples{
data(mtcars)
mtcars.DT <- as.data.table(mtcars)
mtcars.DT \%>\%
 .[, cyl := as.character(cyl)] \%>\%
 pipe_test(test = class(.[["cyl"]]), value = "character")

}
