% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_blocks.R
\name{split_blocks}
\alias{split_blocks}
\title{To create the blocks for bootstrapping}
\usage{
split_blocks(x, m)
}
\arguments{
\item{x}{vector to split}

\item{m}{the block length}
}
\value{
a list of blocks, where each element of the list has length \code{m}
}
\description{
A function to create the blocks for block bootstrapping
}
\examples{
x <- 1:10
split_blocks(x, 2)
split_blocks(x, 3)
split_blocks(x, 4)
}
