% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_summarize.R
\name{summarizing}
\alias{summarizing}
\alias{DT_summarize}
\alias{DT_mean}
\alias{DT_sd}
\alias{DT_var}
\alias{DT_sum}
\alias{DT_log_diff}
\alias{DT_perc_diff}
\title{Convenient functions for summarizing a \code{data.table}.}
\usage{
DT_summarize(DT, fun, vars, names.extra = NULL)

DT_mean(DT, vars, names.extra = ".mean", na.rm = FALSE)

DT_sd(DT, vars, names.extra = ".sd", na.rm = FALSE)

DT_var(DT, vars, names.extra = ".var", na.rm = FALSE)

DT_sum(DT, vars, names.extra = ".sum", na.rm = FALSE)

DT_log_diff(DT, vars, names.extra = ".log.diff")

DT_perc_diff(DT, vars, names.extra = ".perc.diff")
}
\arguments{
\item{DT}{a \code{data.table}. Can also be \code{.SD} if used
inside another \code{data.table}. see examples below}

\item{fun}{a function used to summarize the data}

\item{vars}{a character vector with variable names}

\item{names.extra}{a string used as a suffix for new variable names}

\item{na.rm}{set to \code{TRUE} to remove \code{NA}s. Default is
\code{FALSE}}
}
\value{
a summarized \code{data.table}
}
\description{
Convenient functions for summarizing a \code{data.table}. These
functions are convenient to use with \code{data.table}'s \code{.SD}
argument Shorthand functions are also available that only require a
\code{data.table}, a character vector of variables, \code{vars},
and an optional \code{names.extra} argument:
}
\details{
\describe{
   \item{DT_mean}{\code{fun = mean} and \code{names.extra} defaults to \code{".mean"}}
   \item{DT_sd}{\code{fun = sd}and and \code{names.extra} defaults to \code{".sd"}}
   \item{DT_var}{\code{fun = var} and \code{names.extra} defaults to \code{".var"}}
   \item{DT_sum}{\code{fun = sum} and \code{names.extra} defaults to \code{".sum"}}
   \item{DT_log_diff}{\code{fun = function(x) log(x[length(x)]) -log(x[1])} and \code{names.extra} defaults to \code{".log.diff"}}
   \item{DT_perc_diff}{\code{fun = function(x) function(x) (x[length(x)]-x[1])/x[1]} and \code{names.extra} defaults to \code{".perc.diff"}}
}
}
\examples{
data(mtcars)
setDT(mtcars) ##Convert to a data.table

##Base use of DT_summarize()
DT_summarize(mtcars, fun = mean, vars = c("mpg","hp"))
##using with by and .SD
mtcars \%>\%
    .[, DT_summarize(.SD, fun = mean, vars = c("mpg","hp")), by = cyl]
##Using the convenience function DT_mean (and leaving the names.extra
##argument as the default)
DT_mean(mtcars, vars = c("mpg", "hp"))
mtcars \%>\%
    .[, DT_mean(.SD, vars = c("mpg","hp")), by = cyl]
##Take the mean of of hp and mpg, and the variance of disp and wt
##Note, we concatenate usign c()
mtcars \%>\%
    .[, c(DT_mean(.SD, vars = c("mpg", "hp")),
          DT_var(.SD, vars = c("disp", "wt"))
          )]
##by cyl
mtcars \%>\%
    .[, c(DT_mean(.SD, vars = c("mpg", "hp")),
          DT_var(.SD, vars = c("disp", "wt"))
          ), by = cyl]
##The mean, standard deviation, and variance of mpg and hp
summary.vars <- c("mpg", "hp")
mtcars \%>\%
    .[, c(DT_mean(.SD, vars = summary.vars),
          DT_sd(.SD, vars = summary.vars),
          DT_var(.SD, vars = summary.vars)), by = cyl]

## Other useful summary functions (may not have a
##useful interpretation here)
summary.vars <- c("mpg", "hp")
mtcars \%>\%
    .[, c(DT_sum(.SD, vars = summary.vars),
          DT_log_diff(.SD, vars = summary.vars),
          DT_perc_diff(.SD, vars = summary.vars)),
      by = cyl]
}
