% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_themes.R
\name{theme_bw_cl}
\alias{theme_bw_cl}
\title{ggplot2 theme based on \code{theme_bw} and \code{theme_cowplot}
that works well with facets}
\usage{
theme_bw_cl(font.size = 11)
}
\arguments{
\item{font.size}{the main font size for the graph. Equivalent to
the \code{font_size} parameter in \code{theme_bw}}
}
\description{
Modified version of \code{theme_bw} and \code{theme_cowplot} for
publication quality graph creation that works well with facets
}
\examples{
library(ggplot2)
data(mtcars)
}
