% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_themes.R
\name{theme_cowplot_cl}
\alias{theme_cowplot_cl}
\title{ggplot2 theme based on \code{theme_cowplot}}
\usage{
theme_cowplot_cl(font.size = 11)
}
\arguments{
\item{font.size}{the main font size for the graph. Equivalent to
the \code{font_size} parameter in \code{theme_cowplot}}
}
\description{
Modified version of \code{theme_cowplot} for publication quality
graph creation.
}
\examples{
library(ggplot2)
data(mtcars)
ggplot(mtcars, aes(x = wt, y = mpg)) +
 geom_point(aes(color = hp)) +
 theme_cowplot_cl() +
 labs(title = "Graph title",
      subtitle = "Graph Subtitle", 
      caption = "Graph caption")
}
