% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cols_by_ref.R
\name{select_by_ref}
\alias{select_by_ref}
\title{Select columns in a data.table by reference}
\usage{
select_by_ref(DT, keep.cols)
}
\arguments{
\item{DT}{a \code{data.table}}

\item{keep.cols}{a character vector of columns in \code{DT}}
}
\description{
Select columns in a data.table by reference
}
\examples{
data(mtcars)
DT <- as.data.table(mtcars)
DT <- select_by_ref(DT, "hp")
head(DT)

}
